LPS-77456 Inline WikiEngineTracker.
